Add history category (with zombies) edit.

+++ code/browser/js/IndexedList.js
@@ -3,21 +3,23 @@ class IndexedList {
     this._index = index;
     this._list  = list;
     this._head  = head;
-    this._view  = new IndexedListView (head != null);
+    this._view  = new IndexedListView (head != null, index != null);
   }
 
   delete() {
-    this._index .delete();
+    if (this._index)
+      this._index .delete();
     this._list  .delete();
     this._view  .remove();
   }
 
-  addHtml (toHtml) {
+  async addHtml (toHtml) {
     this._view  .addHtml (toHtml);
-    this._index .addHtml (this._view .getIndexHtml());
+    if (this._index)
+      await this._index .addHtml (this._view .getIndexHtml());
     if (this._head)
       this._head .addHtml (this._view .getHeadHtml());
-    this._list  .addHtml (this._view .getListHtml());
+    await this._list  .addHtml (this._view .getListHtml());
   }
 
   getList() {
